refactor: reorganize headers into include subdirectories
- Moved internal shared headers to src/include/
- Moved builtin headers to src/commands/builtins/include/
- Updated CMakeLists.txt include_directories accordingly
